+++ cube/impact.cpp
@@ -294,8 +294,8 @@ namespace cube
 
       host_to_client()(point);
 
-      m_pengine->m_pinput->m_mousestate.m_position.x = point.x();
-      m_pengine->m_pinput->m_mousestate.m_position.y = point.y();
+      m_pengine->m_pinput->m_mousestate.m_position.x = (float) point.x();
+      m_pengine->m_pinput->m_mousestate.m_position.y = (float) point.y();
       m_pengine->m_pinput->m_mousestate.m_buttons.left = true;
 
       set_mouse_capture();
@@ -316,8 +316,8 @@ namespace cube
 
       host_to_client()(point);
 
-      m_pengine->m_pinput->m_mousestate.m_position.x = point.x();
-      m_pengine->m_pinput->m_mousestate.m_position.y = point.y();
+      m_pengine->m_pinput->m_mousestate.m_position.x = (float) point.x();
+      m_pengine->m_pinput->m_mousestate.m_position.y = (float) point.y();
       m_pengine->m_pinput->m_mousestate.m_buttons.left = false;
